ENH: Also use the fallback's decision when the SUBJECT's decision has expired
Not just when no decision at all had been recorded for SUBJECT.
The user may express a short-lived decision for individual SUBJECT(s), but also state a forever decision for the entire GROUP. The individual decision should then only persist for the short time, and then fall back to the group decision. So far, the user was re-queried again, because the individual decision had expired.