memoizeDecision: ENH: Allow querying for a recorded decision with --check|--has-decision
This is useful for setup-* scripts that offer multiple options and don't want to offer one again if the user persistently declined the offer. A simple use of offerTo (without persistence) would cause a --check invocation there to fail.
The exit status is chosen so that the inclusion of the --check arg will enter the setup code (i.e. when checking having no answer is the actionable branch), and inside the script then fails (meaning a necessary action (and if it's only the decline of the action) is required).
Add a --has-decision flag that yields the same information, but with inverted exit status; i.e. it fails when no decision has yet been recorded. (This dichotomy is similar to setup-software's --check vs. --dry-run.)

,----[ example setup code ]----
| obtainDecision() { memoizeDecision "${checkArg[@]}" --memoize-group setup-FILE "$@"; }
|
| if [ ! -e "$FILE" ] \
|     && obtainDecision --subject File --verb is --state missing --action 'add it'
| then
|     [ "$isCheck" ] && exit 1
|     $SUDO touch "$FILE"
| fi
`----